figuring out a a good way to do this
